DTNN-1286: Properly align STOR HK packet and use mem_free to calculate memory in use

+++ bpa/stor/inc/bplib_stor.h
@@ -62,10 +62,10 @@ typedef struct BPLib_StorageHkTlm_Payload BPLib_StorageHkTlm_Payload_t;
 
 struct BPLib_StorageHkTlm_Payload
 {
-    size_t   BytesMemInUse;     /** \brief Bytes in memory that are in use */
-    size_t   BytesMemFree;      /** \brief Number of bytes free */
-    uint32_t BytesMemHighWater; /** \brief Memory high water mark in bytes */
-    size_t   KbStorageInUse;    /** \brief Kilobytes of storage currently occupied by bundles */
+    size_t BytesMemInUse;     /** \brief Bytes in memory that are in use */
+    size_t BytesMemFree;      /** \brief Number of bytes free */
+    size_t BytesMemHighWater; /** \brief Memory high water mark in bytes */
+    size_t KbStorageInUse;    /** \brief Kilobytes of storage currently occupied by bundles */
 
     int64_t  MonotonicTime;     /** \brief Monotonic Time Counter */
     int64_t  CorrelationFactor; /** \brief Time Correlation Factor */
